For reasons that are a mystery to me, the changes from Guido to vm_mmap never made it into the source tree. From Guido van Rooij:
when mmapping a file, permissions are checked as it should be. When
mprotect()-ing the address range afterwards, no protection is checked
regarding the protection of the file originally opened. So
when you open /usr/bin/su RDONLY and SHARED you can afterwards change
the mmapped region to READ|WRITE. This gives the possibility to obtain
root privs obviously.