Make this compile again after the syscons/psm/kbdio changes that removed the keyboard IO port definitions from <machine/console.h>.
Note: I do not have a machine available to test this yet,  I sent this
for review to several people who have not replied.

This is also totally busted in 2.2, which has half of the syscons/psm/kbdio
changes in it.  This really desperately needs to be tested and fixed
in 2.2 as well.