MFC revision 1.22: When size is 1 should just null terminate the string.
This fix has the side effect of making vsnprintf() properly return the
length of the string that would have been created if the buffer were
infinite length when the passed size is 1.  Previously, it would
erroniously return 0 which causes auto-extending sbufs to misbehave in
userland apps linked against libsbuf.

Approved by:	re (jhb)