Add comments to deter people from using this file directly.  That's not the intention.
Rearrange "you may want to change these values" values to the top, and
reduce their number as much as possible.

tr macro: Require a parameter (because gdb is too stupid to understand
optional parameters), and create macros tr0, tr1 and trf which call it
to connect to /dev/cuaa0, /dev/cuaa1 and firewire connections
respectively.

Split kld symbol load into two separate macros: revision 1.5
simplified things for the /dev/mem case, but broke it for anything
else.  Now the simple /dev/mem version is called kldsyms, and the
version for serial debugging and processor dumps is called getsyms,
and still requires this irritating cut and paste.

Change comments on startup to make life easier for the poor
(de)bugger.